refactor: move cache under util package
