+++ src/main/java/pl/cleankod/ApplicationInitializer.java
@@ -9,8 +9,8 @@ import org.springframework.boot.SpringBootConfiguration;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.core.env.Environment;
-import pl.cleankod.cache.CacheInMemoryClient;
-import pl.cleankod.cache.inmemory.CacheInMemoryLru;
+import pl.cleankod.util.cache.CacheInMemoryClient;
+import pl.cleankod.util.cache.inmemory.CacheInMemoryLru;
 import pl.cleankod.exchange.core.gateway.AccountRepository;
 import pl.cleankod.exchange.core.gateway.CurrencyConversionService;
 import pl.cleankod.exchange.core.usecase.FindAccountAndConvertCurrencyUseCase;
